+++ src/database/migrations/20200521010938-update-monthly-payments.js
@@ -3,12 +3,12 @@ module.exports = {
     return Promise.all([
       queryInterface.removeColumn('monthly_payments', 'value'),
       queryInterface.addColumn('monthly_payments', 'due_value', {
-        type: Sequelize.FLOAT(2),
+        type: Sequelize.DECIMAL,
         allowNull: false,
         defaultValue: 0,
       }),
       queryInterface.addColumn('monthly_payments', 'paid_value', {
-        type: Sequelize.FLOAT(2),
+        type: Sequelize.DECIMAL,
         allowNull: false,
         defaultValue: 0,
       }),
